Allow :order to be passed into `to_list`
The `reduce_nodes` function allows for an order to be specified, and
`to_list` calls this function but does not allow you to specify the
order that you would like. Here, we allow for a keyword list in
idiomatic elixir style that has an `:order` option.

Ideally, I would like `reduce_nodes` and so also `to_list` to provide
an explanation about the order options - what they mean, etc.

However, that seems beyond the scope of this PR.